v1.2.0

+++ pyqt_openai/gpt_widget/left_sidebar/chatNavWidget.py
@@ -1,21 +1,18 @@
 from PySide6.QtCore import Signal, QSortFilterProxyModel, Qt
 from PySide6.QtSql import QSqlTableModel, QSqlQuery
-from PySide6.QtWidgets import QWidget, QVBoxLayout, QMessageBox, QPushButton, QStyledItemDelegate, QTableView, \
-    QAbstractItemView, \
-    QHBoxLayout, \
+from PySide6.QtWidgets import QWidget, QVBoxLayout, QMessageBox, QPushButton, QStyledItemDelegate, QHBoxLayout, \
     QLabel, QSpacerItem, QSizePolicy, QComboBox, QDialog
 
-from pyqt_openai import THREAD_ORDERBY, ICON_ADD, ICON_DELETE, ICON_IMPORT, ICON_SAVE, ICON_CLOSE, \
-    ICON_REFRESH
+import pyqt_openai.globals
+from pyqt_openai import THREAD_ORDERBY, ICON_ADD, ICON_IMPORT, ICON_SAVE, ICON_REFRESH
 from pyqt_openai.gpt_widget.left_sidebar.chatImportDialog import ChatImportDialog
 from pyqt_openai.gpt_widget.left_sidebar.exportDialog import ExportDialog
 from pyqt_openai.gpt_widget.left_sidebar.importDialog import ImportDialog
 from pyqt_openai.lang.translations import LangClass
 from pyqt_openai.models import ChatThreadContainer
-from pyqt_openai.pyqt_openai_data import DB
+from pyqt_openai.globals import DB
 from pyqt_openai.widgets.baseNavWidget import BaseNavWidget
 from pyqt_openai.widgets.button import Button
-from pyqt_openai.widgets.searchBar import SearchBar
 
 
 class FilterProxyModel(QSortFilterProxyModel):
@@ -93,8 +90,6 @@ class ChatNavWidget(BaseNavWidget):
         self.__importBtn.clicked.connect(self.__import)
         self.__saveBtn.clicked.connect(self.__export)
         self.__refreshBtn.clicked.connect(self.__refresh)
-
-        self._delBtn.clicked.connect(self._delete)
 
         lay = QHBoxLayout()
         lay.addWidget(imageGenerationHistoryLbl)
@@ -204,12 +199,17 @@ class ChatNavWidget(BaseNavWidget):
         ids = list(set(ids))
         return ids
 
+    # TODO LANGUAGE
     def _delete(self):
-        ids = self.__getSelectedIds()
-        for _id in ids:
-            DB.deleteThread(_id)
-        self._model.select()
-        self.cleared.emit()
+        reply = QMessageBox.question(self, LangClass.TRANSLATIONS['Confirm'],
+                                     LangClass.TRANSLATIONS['Are you sure to delete the selected data?'],
+                                     QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No)
+        if reply == QMessageBox.StandardButton.Yes:
+            ids = self.__getSelectedIds()
+            for _id in ids:
+                DB.deleteThread(_id)
+            self._model.select()
+            self.cleared.emit()
 
     def _clear(self, table_type='chat'):
         table_type = table_type or 'chat'
